requirements.txt
torch>=1.8.0
torchvision>=0.9.0
mmcv-full>=1.3.0
mmsegmentation>=0.18.0
numpy
opencv-python
pillow
scipy
tqdm